Add files via upload
Students of the college should register in the official portal of college library with their college credentials like registration number, branch, year, etc…After successful completion of registration, they can log in whenever they want and can search for the book they needed and can request for the allocation of that book. If the book is available, it will be allocated for the student and then she/he should go and collect it from the library else she/he should wait for the book until it is available.